@@ -1,6 +1,6 @@
 {
   "name": "@aws-sdk/credential-provider-cognito-identity",
-  "version": "3.259.0",
+  "version": "3.370.0",
   "scripts": {
     "build": "concurrently 'yarn:build:cjs' 'yarn:build:es' 'yarn:build:types'",
     "build:cjs": "tsc -p tsconfig.cjs.json",
@@ -21,10 +21,11 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@aws-sdk/client-cognito-identity": "3.259.0",
-    "@aws-sdk/property-provider": "3.257.0",
-    "@aws-sdk/types": "3.257.0",
-    "tslib": "^2.3.1"
+    "@aws-sdk/client-cognito-identity": "3.370.0",
+    "@aws-sdk/types": "3.370.0",
+    "@smithy/property-provider": "^1.0.1",
+    "@smithy/types": "^1.1.0",
+    "tslib": "^2.5.0"
   },
   "engines": {
     "node": ">=14.0.0"
@@ -37,7 +38,7 @@
     }
   },
   "files": [
-    "dist-*"
+    "dist-*/**"
   ],
   "homepage": "https://github.com/aws/aws-sdk-js-v3/tree/main/packages/credential-provider-cognito-identity",
   "repository": {
@@ -50,7 +51,10 @@
     "concurrently": "7.0.0",
     "downlevel-dts": "0.10.1",
     "rimraf": "3.0.2",
-    "typedoc": "0.19.2",
-    "typescript": "~4.6.2"
+    "typedoc": "0.23.23",
+    "typescript": "~4.9.5"
+  },
+  "typedoc": {
+    "entryPoint": "src/index.ts"
   }
 }
