@@ -1,12 +1,8 @@
 import dotenv from 'dotenv';
 dotenv.config();
-
 import express, { Request, Response } from 'express';
 import bodyParser from 'body-parser';
 import mongoose from 'mongoose';
-import multer from 'multer';
-import { createNotification, deleteNotification, getAllNotifications } from './controllers/notification/notification.controller';
-import { onesignal } from './controllers/notification/oneSignalNotification.controller';
 import cors from 'cors';
 import auth_route from './routes/auth_route';
 import routine_route from './routes/routine_routes';
@@ -15,7 +11,6 @@ import summary from './routes/summary_route';
 import account from './routes/account_route';
 import notice from './routes/notice_route';
 import notification from './routes/notice_route';
-
 const app = express();
 
 // Middleware
@@ -24,28 +19,37 @@ app.use(bodyParser.urlencoded({ extended: true }));
 app.use(cors());
 
 // Connection
-mongoose.connect('mongodb+srv://nurapp:rr1234@cluster0.wwfxxwu.mongodb.net/?retryWrites=true&w=majority')
-  .then(() => console.log('Connected!'));
+const mongodbUri_Test = process.env.MONGODB_URI_TEST_PROJECT || '';
+const mongodbUri_Production = process.env.MONGODB_URI_PRODUCTION_PROJECT || '';
 
-// Routes
+// Connect to the test project
+mongoose.connect(mongodbUri_Test)
+  .then(() => console.log('Connected!'))
+  .catch((err) => console.error('Error connecting to MongoDB Test:', err));
+
+
+
+//****************************************************************************/
+//
+//...............................  Routes.....................................//
+//
+//****************************************************************************/
+//   Account and auth 
 app.use("/auth", auth_route);
+app.use("/account", account);
+
+// Routine 
 app.use("/rutin", routine_route);
 app.use("/class", class_route);
 app.use("/summary", summary);
-app.use("/account", account);
+// NoticeBoard
 app.use("/notice", notice);
+// Notification
 app.use("/notification", notification);
 
-// Multer setup
-const upload = multer({
-  storage: multer.memoryStorage(),
-});
 
-app.post("/notification", upload.single('image'), createNotification);
-app.patch("/notification/:notificationId", deleteNotification);
-app.get("/notification/", getAllNotifications);
 
-app.get("/oneSignal", onesignal);
+// Basic Routes
 
 app.get("/", (req: Request, res: Response) => {
   console.log(req.body);
